python-sdk/pyproject.toml
[project]
name = "python-sdk"
version = "0.1.0"
requires-python = ">=3.11"
readme = "README.md"
authors = [
    { name = "Matteo Manzi" },
]

dependencies = [
    "lighthouseweb3>=0.1.4,<1.0.0",
    "python-dotenv>=1.1.0,<2.0.0",
    "tenseal>=0.3.16,<1.0.0",
    "numpy>=1.26.0,<2.0.0",
    "typer>=0.16.0,<1.0.0",
    "web3>=7.12.0,<8.0.0",
    "pyarrow>=19.0.1,<20.0.0",
    "pandas>=2.1.0,<2.3.0"
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.1.0,<4.2.0",
    "ruff>=0.9.10, < 1.0.0",
    "pydocstyle>=6.3.0,<6.4.0",
    "pytest>=8.3.5,<8.4.0",
    "pydeps>=3.0.0,<4.0.0"
]

[project.scripts]
sdk = "python_sdk.__main__:app"
